requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.6.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.26
langchain-text-splitters==0.2.2
langsmith==0.1.95
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
postgrest==0.16.9
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
storage3==0.7.7
streamlit==1.37.0
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
utils
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
websockets==12.0
yarl==1.9.4
langchain-openai
python-dotenv
utils
streamlit-authenticator
pypdf
langchainhub